+++ .bowerrc
@@ -1,3 +1,3 @@
 {
-  "directory": "client/lib/"
+  "directory": "src/lib/"
 }
