+++ .gitignore
@@ -2,11 +2,10 @@
 .env
 
 # Files that might appear on external disk
+build/
 node_modules/
 bower_components/
 data/db/
-client/lib
-client/build
 
 # Thumbnails
 ._*
