--- client/app/components/boot.jsx
@@ -1,2 +1,0 @@
-React.render(<App clientId={AUTH0_CLIENT_ID} domain={AUTH0_DOMAIN} />,
-  document.getElementById('login-page'));
