--- client/config/auth0-variables.js
@@ -1,3 +1,0 @@
-var AUTH0_CLIENT_ID='cN9g9XPL1kOQKmIgHNZclGgudFQFkFrK'; 
-var AUTH0_DOMAIN='therealericle.auth0.com'; 
-var AUTH0_CALLBACK_URL=location.href;
