+++ package.json
@@ -28,7 +28,6 @@
     "fs": "latest",
     "mongoose": "^4.1.3",
     "morgan": "^1.6.1",
-    "fs": "latest",
     "request": "^2.60.0",
     "flux": "^2.0.1",
     "react": "^0.13.2",
@@ -40,6 +39,7 @@
     "babel-core": "^5.8.23",
     "babel-loader": "^5.3.2",
     "chai": "^3.2.0",
+    "jsx-loader": "^0.13.2",
     "mocha": "^2.2.5",
     "react": "^0.13.3",
     "react-hot-loader": "^1.3.0",
@@ -52,6 +52,9 @@
   "scripts": {
     "test": "npm run test-node",
     "test-node": "mocha -R spec ./test/spec/userModelSpec.js",
-    "start": "webpack --progress --colors --watch -d"
+    "build": "webpack --progress --profile --colors",
+    "watch": "webpack-dev-server --hot --inline --progress --colors",
+    "start": "webpack --progress --colors --watch -d",
+    "dev": "webpack-dev-server --devtool eval --hot --progress --colors --content-base build"
   }
 }
