+++ src/app/components/home.jsx
@@ -1,3 +1,5 @@
+var React = require('react');
+
 var Home = React.createClass({
   showLock: function() {
     this.props.lock.show();
@@ -12,3 +14,5 @@ var Home = React.createClass({
     </div>);
   }
 });
+
+module.exports = Home;
