+++ src/app/components/thumbnail.jsx
@@ -1,0 +1,12 @@
+var React = require('react');
+
+var Thumbnail = React.createClass({
+
+  render: function(){
+    return (
+        <img className="book-img" src={this.props.book.MediumImage.URL}/>
+      );
+  }
+});
+
+module.exports = Thumbnail;
