+++ src/config/auth0.js
@@ -1,0 +1,5 @@
+module.exports = {
+  AUTH0_CLIENT_ID: 'cN9g9XPL1kOQKmIgHNZclGgudFQFkFrK',
+  AUTH0_DOMAIN:'therealericle.auth0.com',
+  AUTH0_CALLBACK_URL: location.href
+};
