+++ src/detail-view.html
