+++ src/index.html
@@ -2,7 +2,10 @@
 <html>
     <head>
         <meta charset="utf-8">
+
+        <!-- jQuery -->
         <script src="http://code.jquery.com/jquery-2.1.1.min.js"></script>
+        <!-- Auth0 -->
         <script src="https://cdn.auth0.com/js/lock-7.js"></script>
 
         <script src="//use.typekit.net/iws6ohy.js"></script>
@@ -13,20 +16,20 @@
         <!-- font awesome from BootstrapCDN -->
         <link href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css" rel="stylesheet">
         <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css" rel="stylesheet">
-        <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.3.15/slick.css" />
 
         <!-- React libs -->
         <!-- The core React library -->
         <script src="http://fb.me/react-0.13.1.js"></script>
         <!-- In-browser JSX transformer, remove when pre-compiling JSX. -->
         <script src="http://fb.me/JSXTransformer-0.13.1.js"></script>
-
         <link href="css/styles.css" rel="stylesheet">
     </head>
     <body class="home">
-        <div id="recommended-gifts">
+        <div class="container">
+            <div class="login-page clearfix" id="login-page"></div>
+            <div id='friend-manager-container'></div>
         </div>
         <script type="text/javascript" src="build/common.js"> </script>
-        <script type="text/javascript" src="build/recommendedGifts.js"> </script>
+        <script type="text/javascript" src="build/bundle.js"> </script>
     </body>
 </html>
