+++ src/recommended.html
@@ -2,12 +2,9 @@
 <html>
     <head>
         <meta charset="utf-8">
-
-        <!-- jQuery -->
         <script src="http://code.jquery.com/jquery-2.1.1.min.js"></script>
-        <!-- Auth0 -->
         <script src="https://cdn.auth0.com/js/lock-7.js"></script>
-        
+
         <script src="//use.typekit.net/iws6ohy.js"></script>
         <script>try{Typekit.load();}catch(e){}</script>
 
@@ -16,6 +13,7 @@
         <!-- font awesome from BootstrapCDN -->
         <link href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css" rel="stylesheet">
         <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css" rel="stylesheet">
+        <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.3.15/slick.css" />
 
         <!-- React libs -->
         <!-- The core React library -->
@@ -23,18 +21,12 @@
         <!-- In-browser JSX transformer, remove when pre-compiling JSX. -->
         <script src="http://fb.me/JSXTransformer-0.13.1.js"></script>
 
-        <script src="config/auth0-variables.js"> </script>
-        <script type="text/jsx" src="app/components/loggedIn.jsx"> </script>
-        <script type="text/jsx" src="app/components/home.jsx"> </script>
-        <script type="text/jsx" src="app/components/app.jsx"> </script>
-        <script type="text/jsx" src="app/components/boot.jsx"> </script>
-        <script type="text/jsx" src="app/components/friendManager.jsx"> </script>
         <link href="css/styles.css" rel="stylesheet">
     </head>
     <body class="home">
-        <div class="container">
-            <div class="login-page clearfix" id="login-page"></div>
-            <div id='friend-manager-container'></div>
+        <div id="recommended-gifts">
         </div>
+        <script type="text/javascript" src="build/common.js"> </script>
+        <script type="text/javascript" src="build/bundle.js"> </script>
     </body>
 </html>
